requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
attrs==25.3.0
autobahn==24.4.2
Automat==25.4.16
bidict==0.23.1
certifi==2025.7.9
cffi==1.17.1
channels==4.2.2
channels_redis==4.2.1
click==8.2.1
constantly==23.10.4
cryptography==45.0.5
daphne==4.2.1
Django==5.2.3
django-cors-headers==4.7.0
django-filter==25.1
django-htmx==1.23.2
django-jazzmin==3.0.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
dnspython==2.7.0
docutils==0.21.2
drf-spectacular==0.28.0
drf-yasg==1.21.10
fastapi==0.116.1
frozenlist==1.7.0
greenlet==3.2.3
h11==0.16.0
htmx==0.0.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
hyperlink==21.0.0
idna==3.10
ifaddr==0.2.0
incremental==24.7.2
inflection==0.5.1
itsdangerous==2.2.0
Jinja2==3.1.6
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
markdown2==2.5.3
MarkupSafe==3.0.2
msgpack==1.1.1
multidict==6.6.3
nicegui==2.21.1
orjson==3.10.18
packaging==25.0
pillow==11.2.1
propcache==0.3.2
pscript==0.7.7
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==1.10.22
pydantic_core==2.33.2
Pygments==2.19.2
PyJWT==2.9.0
pymemcache==4.0.0
pymongo==4.13.2
pyOpenSSL==25.1.0
python-dotenv==1.1.1
python-engineio==4.12.2
python-multipart==0.0.20
python-socketio==5.13.0
pytz==2025.2
PyYAML==6.0.2
redis==6.2.0
referencing==0.36.2
rpds-py==0.26.0
service-identity==24.2.0
setuptools==80.9.0
shortuuid==1.0.13
simple-websocket==1.1.0
sniffio==1.3.1
SQLAlchemy==2.0.41
sqlparse==0.5.3
starlette==0.47.1
Twisted==25.5.0
txaio==25.6.1
typing-inspection==0.4.1
typing_extensions==4.14.0
uritemplate==4.2.0
uvicorn==0.35.0
uvloop==0.21.0
vbuild==0.8.2
watchfiles==1.1.0
websockets==15.0.1
wsproto==1.2.0
yarl==1.20.1
zope.interface==7.2
